pyproject.toml
[tool.poetry]
name = "picard-epo"
version = "0.1.0"
description = ""
authors = ["Andrew Zheng <azheng92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mpax = "^0.2.4"
pyepo = "^1.0.2"
scikit-learn = "^1.7.2"
wandb = "^0.22.0"
jax-tqdm = "^0.4.0"
hydra-core = "^1.3.2"
pooch = "^1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
